--- kptracker_serv/kpbt/bowling_centers/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import render, redirect
-from django.contrib.auth.decorators import login_required
-
-def create_bowling_center(request):
-	return render(request, 'bowling_centers/create.html', context={})
