+++ kptracker_serv/kpbt/centers/templates/centers/create_center.html
@@ -1,0 +1,12 @@
+{% extends 'base_generic.html' %}
+
+{% block title %} Signup {% endblock %}
+
+{% block content %}
+	<h2>Create Bowling Center</h2>
+	<form method="post">
+		{% csrf_token %}
+		{{ form.as_p }}
+		<button type="submit">Create </button>
+	</form>
+{% endblock %}
