+++ kptracker_serv/kpbt/centers/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from . import views
+from kpbt.centers import views
 
 urlpatterns =[
 	path('create/', views.create_bowling_center, name='create-bowling-center')
